Fixing code and build configuration
