+++ pom.xml
@@ -90,6 +90,7 @@
 
         <!-- Maven Plugin Versions -->
         <maven-compiler-plugin.version>3.5</maven-compiler-plugin.version>
+        <maven-deploy-plugin.version>2.8.2</maven-deploy-plugin.version>
         <maven-surefire-plugin.version>2.19.1</maven-surefire-plugin.version>
         <fabric8.maven.plugin.version>3.1.69</fabric8.maven.plugin.version>
         <gmavenplus-plugin.version>1.2</gmavenplus-plugin.version>
