+++ spring-cloud-kubernetes-core/src/main/java/io/fabric8/spring/cloud/kubernetes/reload/EventBasedConfigurationChangeDetector.java
@@ -1,11 +1,15 @@
 package io.fabric8.spring.cloud.kubernetes.reload;
 
+import java.util.HashMap;
+import java.util.Map;
 import javax.annotation.PostConstruct;
+import javax.annotation.PreDestroy;
 
 import io.fabric8.kubernetes.api.model.ConfigMap;
 import io.fabric8.kubernetes.api.model.Secret;
 import io.fabric8.kubernetes.client.KubernetesClient;
 import io.fabric8.kubernetes.client.KubernetesClientException;
+import io.fabric8.kubernetes.client.Watch;
 import io.fabric8.kubernetes.client.Watcher;
 import io.fabric8.spring.cloud.kubernetes.config.ConfigMapPropertySource;
 import io.fabric8.spring.cloud.kubernetes.config.ConfigMapPropertySourceLocator;
@@ -24,27 +28,27 @@ public class EventBasedConfigurationChangeDetector extends ConfigurationChangeDe
 
     private SecretsPropertySourceLocator secretsPropertySourceLocator;
 
-    private EventWatcher eventWatcher;
+    private Map<String, Watch> watches;
 
     public EventBasedConfigurationChangeDetector(AbstractEnvironment environment,
                                                  ConfigReloadProperties properties,
                                                  KubernetesClient kubernetesClient,
                                                  ConfigurationUpdateStrategy strategy,
                                                  ConfigMapPropertySourceLocator configMapPropertySourceLocator,
-                                                 SecretsPropertySourceLocator secretsPropertySourceLocator,
-                                                 EventWatcher eventWatcher) {
+                                                 SecretsPropertySourceLocator secretsPropertySourceLocator) {
         super(environment, properties, kubernetesClient, strategy);
 
-        this.eventWatcher = eventWatcher;
         this.configMapPropertySourceLocator = configMapPropertySourceLocator;
         this.secretsPropertySourceLocator = secretsPropertySourceLocator;
+        this.watches = new HashMap<>();
     }
 
     @PostConstruct
     public void watch() {
 
         if (properties.isMonitoringConfigMaps()) {
-            eventWatcher.addWatch("config-maps-watch", k -> k.configMaps()
+            String name = "config-maps-watch";
+            watches.put(name, kubernetesClient.configMaps()
                     .watch(new Watcher<ConfigMap>() {
                         @Override
                         public void eventReceived(Action action, ConfigMap configMap) {
@@ -55,10 +59,12 @@ public class EventBasedConfigurationChangeDetector extends ConfigurationChangeDe
                         public void onClose(KubernetesClientException e) {
                         }
                     }));
+            log.info("Added new Kubernetes watch: {}", name);
         }
 
         if (properties.isMonitoringSecrets()) {
-            eventWatcher.addWatch("secrets-watch", k -> k.secrets()
+            String name = "secrets-watch";
+            watches.put(name, kubernetesClient.secrets()
                     .watch(new Watcher<Secret>() {
                         @Override
                         public void eventReceived(Action action, Secret secret) {
@@ -69,9 +75,25 @@ public class EventBasedConfigurationChangeDetector extends ConfigurationChangeDe
                         public void onClose(KubernetesClientException e) {
                         }
                     }));
+            log.info("Added new Kubernetes watch: {}", name);
         }
 
         log.info("Kubernetes polling configuration change detector activated");
+    }
+
+    @PreDestroy
+    public void unwatch() {
+        if (this.watches != null) {
+            for (Map.Entry<String, Watch> entry : this.watches.entrySet()) {
+                try {
+                    log.debug("Closing the watch {}", entry.getKey());
+                    entry.getValue().close();
+
+                } catch (Exception e) {
+                    log.error("Error while closing the watch connection", e);
+                }
+            }
+        }
     }
 
     private void onEvent(ConfigMap configMap) {
